You can now calculate heart rate data :)

+++ cycleFitMAUI/Data/CreateWorkout.cs
@@ -198,7 +198,7 @@ namespace cycleFitMAUI.Data
 
     public class Zone
     {
-        public int Low { get; set; }
-        public int High { get; set; }
+        public double Low { get; set; }
+        public double High { get; set; }
     }
 }
